+++ HISTORY.md
@@ -8,6 +8,8 @@ v0.89-develop
    * dcb9b4644f fix standalone detection on single `{{.}}` or `{{this}}`
    * 2bbd50bd1f support `{{#if foo includeZero=true}}`
    * b0b4e51427 support literal references
+   * **BREAK CHANGE** remove FLAG_MUSTACHEPAIN
+   * new flag FLAG_PREVENTINDENT to stop auto indent on partial.
 
 v0.23 https://github.com/zordius/lightncandy/tree/v0.23
    * align with handlebars.js 3.0.3